Some various updates and bug fixes

+++ mohea/resources/js/listing/Listing.js
@@ -8,6 +8,10 @@ const Listing = () => {
   const [listing, setListing] = useState([]);
   const [sortBy, setSortBy] = useState('');
   let isDeleting = false;
+  let isCanceling = false;
+  let isError = false;
+  let currentIdx = "";
+  let deleteTimeout;
 
   useEffect(() => {
     axios({
@@ -21,34 +25,64 @@ const Listing = () => {
       })
   }, [])
 
-  const handleDelete = (id) => {
-    isDeleting = true;
-    axios({
-      method: 'DELETE',
-      url: url + '/project/' + id,
-    })
-      .then(res => {
-        if (res.status === 200) {
-          growl({
-            type: 'success',
-            message: <b>Your project has been deleted</b>
-          });
+  const handleDelete = (id, idx) => {
+    if (currentIdx == "") {
+      currentIdx = idx;
+    }
+    if (currentIdx === idx) {
+      isDeleting = !isDeleting;
+      document.getElementById('listing-' + idx).classList.toggle("deleting");
+      if (isDeleting) {
+        document.getElementById('button-' + idx).innerHTML = "Cancel";
+        isCanceling = true;
+        deleteTimeout = setTimeout(() => {
           axios({
-            method: 'GET',
-            url: url + '/project/',
+            method: 'DELETE',
+            url: url + '/project/' + id,
           })
             .then(res => {
               if (res.status === 200) {
-                setListing(res.data);
+                growl({
+                  type: 'success',
+                  message: <b>Your project has been deleted</b>
+                });
+                axios({
+                  method: 'GET',
+                  url: url + '/project/',
+                })
+                  .then(res => {
+                    if (res.status === 200) {
+                      setListing(res.data);
+                    }
+                  })
+                isDeleting = false;
               }
             })
-          isDeleting = false;
-        }
-      })
+        }, 5000);
+      }
+      else {
+        clearTimeout(deleteTimeout);
+        document.getElementById('button-' + idx).innerHTML = "Delete";
+        currentIdx = "";
+        setTimeout(() => {
+          isCanceling = false;
+        }, 200);
+      }
+    }
+    else {
+      isError = true;
+      growl({
+        type: 'error',
+        message: <b>An other project is deleting</b>
+      });
+      setTimeout(() => {
+        isError = false;
+      }, 200);
+    }
   }
 
   const handleUpdate = (id, type) => {
-    if (!isDeleting) {
+    if (!isDeleting && !isCanceling && !isError) {
       window.location.href = url + '/project/' + type + '/' + id;
     }
     return;
@@ -70,12 +104,12 @@ const Listing = () => {
         }
 
         return (
-          <button key={'listing-' + idx} onClick={() => handleUpdate(item.id, item.type)} className={item.type + "-item"}>
+          <button key={'listing-' + idx} id={'listing-' + idx} onClick={() => handleUpdate(item.id, item.type)} className={item.type + "-item"}>
             <h4>{item.name ? item.name : 'New ' + item.type}</h4>
             {item.caption &&
               <p>{item.caption}</p>
             }
-            <span className="button" onClick={() => handleDelete(item.id)} tabIndex="0">Delete</span>
+            <span className="button" id={'button-' + idx} onClick={() => handleDelete(item.id, idx)} tabIndex="0">Delete</span>
           </button>
         )
       })}
